// src/app/services/models/datosContacto.ts
// tslint:disable-next-line: class-name
export interface DatosContacto {
    personaContactoEmergencia?: string;
    telefonoContactoEmergencia?: number;
    parentescoContactoEmergencia?: string;
    emailContactoEmergencia?: string;
    cep?: string;
    numCasa?: string;
    logradouroResidencia?: string;
    barrioResidencia?: string;
    ciudadResidencia?: string;
    estadoResidencia?: string;
    complemento?: string;
    email?: string;
    telefono?: string;
}